backend/requirements.txt
fastapi>=0.115.6
uvicorn>=0.33.0
motor>=3.6.0
slowapi>=0.1.9
python-dotenv>=1.0.1
pillow>=10.4.0
pytesseract>=0.3.13
pdf2image>=1.17.0
python-multipart>=0.0.20